// src/app/demo/models/modelsbienraiz/empresabienraizviewmodel.ts
export class EmpresaBienRaiz {
    embr_Id?: number;
    embr_Nombre?: string;
    embr_ContactoA?: string;
    embr_ContactoB?: string;
    embr_TelefonoA?: string;
    embr_TelefonoB?: string;
    usuaCreacion?: string;
    usuaModificacion?: string;
    usua_Creacion?: number;
    embr_FechaCreacion?: string;
    usua_Modificacion?: number;
    embr_FechaModificacion?: string;
    embr_Estado?: string;
    codigo?: string;
}

export class DropDownEmpresas{
    embr_Id?:string;
    embr_Nombre?:string
   }